# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Accounts Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EconomicRecord {
  year: number;
  /** GDP in billions USD */
  gdp: number;
  /** GNP in billions USD */
  gnp: number;
  /** NDP in billions USD */
  ndp: number;
  /** Inflation rate as percentage */
  inflationRate: number;
  /** Unemployment rate as percentage */
  unemploymentRate: number;
  /** Year-over-year GDP growth rate as percentage */
  gdpGrowthRate: number;
}

export interface EconomicSummary {
  latestYear: number;
  gdp: number;
  gnp: number;
  ndp: number;
  inflationRate: number;
  unemploymentRate: number;
  gdpGrowthRate: number;
  /** Change from previous year */
  gdpChange: number;
  inflationChange: number;
  unemploymentChange: number;
}

export interface SectorRecord {
  year: number;
  sector: string;
  /** Contribution value in billions USD */
  value: number;
  /** Percentage of total GDP */
  percentage: number;
}

export interface EconomicTrends {
  gdpTrend: EconomicRecord[];
  inflationTrend: EconomicRecord[];
  sectorTrend: SectorRecord[];
  availableYears: number[];
}

export type GetEconomicDataParams = {
  /**
   * Start year filter
   * @nullable
   */
  startYear?: number | null;
  /**
   * End year filter
   * @nullable
   */
  endYear?: number | null;
};

export type GetSectorDataParams = {
  /**
   * Year to get sector data for
   * @nullable
   */
  year?: number | null;
};