# lib/api-zod/src/generated/types/economicRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Accounts Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface EconomicRecord {
  year: number;
  /** GDP in billions USD */
  gdp: number;
  /** GNP in billions USD */
  gnp: number;
  /** NDP in billions USD */
  ndp: number;
  /** Inflation rate as percentage */
  inflationRate: number;
  /** Unemployment rate as percentage */
  unemploymentRate: number;
  /** Year-over-year GDP growth rate as percentage */
  gdpGrowthRate: number;
}