# lib/api-zod/src/generated/types/economicSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Accounts Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface EconomicSummary {
  latestYear: number;
  gdp: number;
  gnp: number;
  ndp: number;
  inflationRate: number;
  unemploymentRate: number;
  gdpGrowthRate: number;
  /** Change from previous year */
  gdpChange: number;
  inflationChange: number;
  unemploymentChange: number;
}