# lib/api-zod/src/generated/types/economicTrends.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Accounts Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EconomicRecord } from "./economicRecord";
import type { SectorRecord } from "./sectorRecord";

export interface EconomicTrends {
  gdpTrend: EconomicRecord[];
  inflationTrend: EconomicRecord[];
  sectorTrend: SectorRecord[];
  availableYears: number[];
}