# lib/api-zod/src/generated/types/sectorRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Accounts Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SectorRecord {
  year: number;
  sector: string;
  /** Contribution value in billions USD */
  value: number;
  /** Percentage of total GDP */
  percentage: number;
}